refactor(capture): photo screen 내부 view와 logic을 분리 [usCameraPermission] refactor: features/capture내부 component로 capture컴포넌트 생성
